+++ AllPrograms/AllPrograms.sln
@@ -1,0 +1,41 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28010.2019
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "TemplateWithClasses", "TemplateWithClasses\TemplateWithClasses.vcxproj", "{8D4B6A5F-B6E7-4DC6-8BDF-137CEA1E34A0}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ExampleWithCodingStyle", "ExampleWithCodingStyle\ExampleWithCodingStyle.vcxproj", "{92BC94AA-451C-4CD7-9A17-0BA2F179A7AF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8D4B6A5F-B6E7-4DC6-8BDF-137CEA1E34A0}.Debug|x64.ActiveCfg = Debug|x64
+		{8D4B6A5F-B6E7-4DC6-8BDF-137CEA1E34A0}.Debug|x64.Build.0 = Debug|x64
+		{8D4B6A5F-B6E7-4DC6-8BDF-137CEA1E34A0}.Debug|x86.ActiveCfg = Debug|Win32
+		{8D4B6A5F-B6E7-4DC6-8BDF-137CEA1E34A0}.Debug|x86.Build.0 = Debug|Win32
+		{8D4B6A5F-B6E7-4DC6-8BDF-137CEA1E34A0}.Release|x64.ActiveCfg = Release|x64
+		{8D4B6A5F-B6E7-4DC6-8BDF-137CEA1E34A0}.Release|x64.Build.0 = Release|x64
+		{8D4B6A5F-B6E7-4DC6-8BDF-137CEA1E34A0}.Release|x86.ActiveCfg = Release|Win32
+		{8D4B6A5F-B6E7-4DC6-8BDF-137CEA1E34A0}.Release|x86.Build.0 = Release|Win32
+		{92BC94AA-451C-4CD7-9A17-0BA2F179A7AF}.Debug|x64.ActiveCfg = Debug|x64
+		{92BC94AA-451C-4CD7-9A17-0BA2F179A7AF}.Debug|x64.Build.0 = Debug|x64
+		{92BC94AA-451C-4CD7-9A17-0BA2F179A7AF}.Debug|x86.ActiveCfg = Debug|Win32
+		{92BC94AA-451C-4CD7-9A17-0BA2F179A7AF}.Debug|x86.Build.0 = Debug|Win32
+		{92BC94AA-451C-4CD7-9A17-0BA2F179A7AF}.Release|x64.ActiveCfg = Release|x64
+		{92BC94AA-451C-4CD7-9A17-0BA2F179A7AF}.Release|x64.Build.0 = Release|x64
+		{92BC94AA-451C-4CD7-9A17-0BA2F179A7AF}.Release|x86.ActiveCfg = Release|Win32
+		{92BC94AA-451C-4CD7-9A17-0BA2F179A7AF}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8B542478-8E5F-45F1-B72F-379029B46E9B}
+	EndGlobalSection
+EndGlobal
