+++ AllPrograms/TemplateWithClasses/TemplateWithClasses/SimpleClass.cpp
@@ -1,0 +1,1 @@
+#include "SimpleClass.h"
