--- TemplateWithClasses.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28010.2019
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "TemplateWithClasses", "TemplateWithClasses\TemplateWithClasses.vcxproj", "{2DD93135-DD70-4859-B389-3DDFE9A49A05}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2DD93135-DD70-4859-B389-3DDFE9A49A05}.Debug|x64.ActiveCfg = Debug|x64
-		{2DD93135-DD70-4859-B389-3DDFE9A49A05}.Debug|x64.Build.0 = Debug|x64
-		{2DD93135-DD70-4859-B389-3DDFE9A49A05}.Debug|x86.ActiveCfg = Debug|Win32
-		{2DD93135-DD70-4859-B389-3DDFE9A49A05}.Debug|x86.Build.0 = Debug|Win32
-		{2DD93135-DD70-4859-B389-3DDFE9A49A05}.Release|x64.ActiveCfg = Release|x64
-		{2DD93135-DD70-4859-B389-3DDFE9A49A05}.Release|x64.Build.0 = Release|x64
-		{2DD93135-DD70-4859-B389-3DDFE9A49A05}.Release|x86.ActiveCfg = Release|Win32
-		{2DD93135-DD70-4859-B389-3DDFE9A49A05}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7EDA935B-B9CC-4CA8-BFBD-B56190F4DCB0}
-	EndGlobalSection
-EndGlobal
