--- TemplateWithClasses/SimpleClass.cpp
@@ -1,1 +1,0 @@
-#include "SimpleClass.h"
