--- TemplateWithClasses/Types.h
@@ -1,7 +1,0 @@
-#ifndef TYPES_H
-#define TYPES_H
-
-typedef int Type;
-typedef double SomeOutputType;
-
-#endif // TYPES_H
